requirements.txt
requests
python-dotenv
streamlit
Pillow

